+++ oggm/tests/test_models.py
@@ -14,6 +14,7 @@ from numpy.testing import assert_allclose
 import pytest
 
 # Local imports
+import oggm
 from oggm.core import massbalance
 from oggm.core.massbalance import LinearMassBalance
 import xarray as xr
@@ -45,12 +46,12 @@ _url_retrieve = None
 
 
 def setup_module(module):
-    module._url_retrieve = utils._urlretrieve
-    utils._urlretrieve = patch_url_retrieve_github
+    module._url_retrieve = utils.oggm_urlretrieve
+    oggm.utils._downloads.oggm_urlretrieve = patch_url_retrieve_github
 
 
 def teardown_module(module):
-    utils._urlretrieve = module._url_retrieve
+    oggm.utils._downloads.oggm_urlretrieve = module._url_retrieve
 
 
 pytestmark = pytest.mark.test_env("models")
