+++ oggm/utils/__init__.py
@@ -1,0 +1,4 @@
+# flake8: noqa
+from oggm.utils._downloads import *
+from oggm.utils._funcs import *
+from oggm.utils._workflow import *
